Refactor Telkomsel Cash package name
